combine files into a single one
