+++ ddtrace/contrib/botocore/patch.py
@@ -2,6 +2,8 @@
 Trace queries to aws api done via botocore client
 """
 # 3p
+import base64
+import json
 from ddtrace.vendor import wrapt
 from ddtrace import config
 import botocore.client
@@ -12,7 +14,8 @@ from ...ext import SpanTypes, http, aws
 from ...pin import Pin
 from ...utils.formats import deep_getattr
 from ...utils.wrappers import unwrap
-from .awslambda import inject_trace_to_client_context
+from ...internal.logger import get_logger
+from ...propagation.http import HTTPPropagator
 from .sqs import inject_trace_to_message_metadata
 
 # Original botocore client class
@@ -20,6 +23,61 @@ _Botocore_client = botocore.client.BaseClient
 
 ARGS_NAME = ('action', 'params', 'path', 'verb')
 TRACED_ARGS = ['params', 'path', 'verb']
+
+log = get_logger(__name__)
+propagator = HTTPPropagator()
+
+
+def inject_trace_to_message_metadata(args, span):
+    trace_headers = {}
+    propagator.inject(span.context, trace_headers)
+    params = args[1]
+
+    if 'MessageAttributes' not in params:
+        params['MessageAttributes'] = {}
+    if len(params['MessageAttributes']) < 10:
+        params['MessageAttributes']['_datadog'] = {
+            'DataType': 'String',
+            'StringValue': json.dumps(trace_headers)
+        }
+
+
+def modify_client_context(client_context_base64, trace_headers):
+    try:
+        client_context_json = base64.b64decode(client_context_base64).decode('utf-8')
+        client_context_object = json.loads(client_context_json)
+
+        if 'custom' in client_context_object:
+            client_context_object['custom']['_datadog'] = trace_headers
+        else:
+            client_context_object['custom'] = {
+                '_datadog': trace_headers
+            }
+
+        new_context = base64.b64encode(json.dumps(client_context_object).encode('utf-8')).decode('utf-8')
+        return new_context
+    except Exception:
+        log.warning('malformed client_context=%s', client_context_base64, exc_info=True)
+        return client_context_base64
+
+
+def inject_trace_to_client_context(args, span):
+    trace_headers = {}
+    propagator.inject(span.context, trace_headers)
+
+    params = args[1]
+    if 'ClientContext' in params:
+        params['ClientContext'] = modify_client_context(params['ClientContext'], trace_headers)
+    else:
+        trace_headers = {}
+        propagator.inject(span.context, trace_headers)
+        client_context_object = {
+            'custom': {
+                '_datadog': trace_headers
+            }
+        }
+        json_context = json.dumps(client_context_object).encode('utf-8')
+        params['ClientContext'] = base64.b64encode(json_context).decode('utf-8')
 
 
 def patch():
